JMC patch

+++ JMCClasses/Core/database/JMCDatabase.m
@@ -942,7 +942,7 @@
 }
 
 - (NSString*)description {
-    return [NSString stringWithFormat:@"%@ %d hit(s) for query %@", [super description], (int)useCount, query];
+    return [NSString stringWithFormat:@"%@ %d hit(s) for query %@", [super description], useCount, query];
 }
